server/requirements.txt
# SmartPark Server Dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# MQTT
paho-mqtt==1.6.1

# ML / Computer Vision
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Geometry (for slot polygon operations)
shapely==2.0.2

# Async file handling
aiofiles==23.2.1
